+++ tests/controller.py
@@ -10,3 +10,8 @@ class Resource(object):
 @Resource.controller('engage')
 def engage(resource, request):
     return {'message': 'Ai ai captain'}
+
+
+@Resource.controller('engage', method='GET')
+def get_engage(resource, request):
+    return {'message': 'Warp engine offline'}